Corrections in glexec test as per Tim's suggestions, also added a separate piece of code to correctly and automatically define the lcmaps.db
git-svn-id: https://vdt.cs.wisc.edu/svn/software/osg-test/trunk@14105 4e558342-562e-0410-864c-e07659590f8c

+++ osgtest/tests/test_44_glexec.py
@@ -12,8 +12,7 @@ class TestGlexec(unittest.TestCase):
 
     # Constants
     __glexec_client_cert = '/tmp/x509_client_cert'
-    __grid_mapfile = '/etc/grid-security/grid-mapfile'
-
+    __grid_mapfile = core.config['system.mapfile']   # typically '/etc/grid-security/grid-mapfile'
 
     # attributes to be filled later
     __uid = ''
@@ -23,6 +22,10 @@ class TestGlexec(unittest.TestCase):
     # ==========================================================================
 
     def test_01_check_gridmap(self):
+        if not core.rpm_is_installed('glexec'):
+            core.skip('not installed')
+            return
+
         pwd_entry = pwd.getpwnam(core.options.username)
         cert_path = os.path.join(pwd_entry.pw_dir, '.globus', 'usercert.pem')
         user_cert_dn, user_cert_issuer = core.certificate_info(cert_path)
@@ -35,24 +38,34 @@ class TestGlexec(unittest.TestCase):
         TestGlexec.__good_gridmap = True
 
     def test_02_define_user_proxy_path(self):
+        if not core.rpm_is_installed('glexec'):
+            core.skip('not installed')
+            return
         command = ('/usr/bin/id','-u')
         status, stdout, stderr = core.system(command, True)
         TestGlexec.__uid = stdout.rstrip()
         TestGlexec.__user_proxy_path = '/tmp/x509up_u'+self.__uid
 
     def test_03_create_user_proxy(self):
-        # hostname = socket.getfqdn()
+        if not core.rpm_is_installed('glexec'):
+            core.skip('not installed')
+            return
 
         command = ('grid-proxy-init','-out',self.__user_proxy_path)
         password = core.options.password + '\n'
         status, stdout, stderr = core.system(command, True, password)
         self.assert_(status==0, 'grid-proxy-init for user '+core.options.username+' has failed')
 
+        # we need to have the right permissions on that proxy for glexec to agree to work,
+        # and the easiest way is to copy the file
         command = ('/bin/cp', self.__user_proxy_path, self.__glexec_client_cert)
         status, stdout, stderr = core.system(command)
         os.environ['GLEXEC_CLIENT_CERT']=self.__glexec_client_cert
 
     def test_04_glexec_switch_id(self):
+        if not core.rpm_is_installed('glexec'):
+            core.skip('not installed')
+            return
         command = ('/usr/sbin/glexec','/usr/bin/id','-u')
 
         status, stdout, stderr = core.system(command)
@@ -61,4 +74,7 @@ class TestGlexec(unittest.TestCase):
         self.assert_(self.__uid==switched_id, 'Glexec identity switch from root to user '+core.options.username+' failed')
 
     def test_05_glexec_proxy_cleanup(self):
+        if not core.rpm_is_installed('glexec'):
+            core.skip('not installed')
+            return
         status = os.unlink(self.__glexec_client_cert)
